// src/lib/types/supabase.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  public: {
    Tables: {
      albums: {
        Row: {
          artist: string;
          artist_id: string;
          cover_url: string | null;
          created_at: string;
          id: string;
          release_date: string | null;
          spotify_id: string;
          spotify_updated_at: string | null;
          title: string;
          tracks: Json;
          updated_at: string;
        };
        Insert: {
          artist: string;
          artist_id: string;
          cover_url?: string | null;
          created_at?: string;
          id?: string;
          release_date?: string | null;
          spotify_id: string;
          spotify_updated_at?: string | null;
          title: string;
          tracks?: Json;
          updated_at?: string;
        };
        Update: {
          artist?: string;
          artist_id?: string;
          cover_url?: string | null;
          created_at?: string;
          id?: string;
          release_date?: string | null;
          spotify_id?: string;
          spotify_updated_at?: string | null;
          title?: string;
          tracks?: Json;
          updated_at?: string;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      rating_value:
        | "0.5"
        | "1.0"
        | "1.5"
        | "2.0"
        | "2.5"
        | "3.0"
        | "3.5"
        | "4.0"
        | "4.5"
        | "5.0";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};